Create test inputs
A message has been generated using MATLAB editor and noise signal has been added to it. The noise function is given as noise = (b-a).*rand(length(sine_wave),1) + a; noise = noise'. To feed the above data to the moving average filter it must be converted into binary data. The at first the real values are converted into integer data type and then it is converted into binary data using the function dec2bin, and the generated data is stored in signal.data file.